requirements.txt
flask==1.0.2
boto3==1.15.18
python-dotenv==0.14.0